+++ icsforge/campaigns/__init__.py
@@ -1,0 +1,1 @@
+"""ICSForge v0.30 Campaign Playbook system."""