+++ icsforge/detections/__init__.py
@@ -1,0 +1,1 @@
+"""ICSForge v0.30 Detection Content Generator."""